refactor(realm): extract node shims from js-realm-shared
